Forgot to switch branches again, I am going to switch rn so I do this again without realizing.
Got Google Maps API requests working. Google maps has a traffic history, so that will be used to set the original wakeup time. However, when the user interacts with the alarm in the morning, the application will make another request to get current traffic, and if there is a lot of traffic it will notify the user and give them a new estimated time to leave.