Added proximity reminder functionality
There was a crash, with the account manager class after a user reminder came in. It was a null pointer exception with the listener.onLoginResponse() method.

I added a null check before calling the method, but I have no idea why it would've been null in the first place, since it was outside the application